function name and module, make kwargs and defaults, docstring
